feat: Faz 2 finansal hesaplama ve tarih dogrulugu
- Teklif KDV'si brüt (matrah) bazlıya çevrildi (iskonto KDV matrahını düşürmez)
- KDV hesaplamaları ortak faturaHesapla.js helper'larına bağlandı (tek VARSAYILAN_KDV_ORANI=20, null-safe)
- Dashboard iadeOrani gerçek satış iadesi / ciro üzerinden hesaplanıyor
- getLocalDateString() merkezi yerel tarih yardımcısı: 36+ toISOString().split('T')[0] deseni taşındı (gece 00:00-02:59 UTC kayması düzeltiliyor)
- Günlük satış görünümleri (HizliSatis, Kasa) yerel tarihle karşılaştırıyor
- YoneticiKokpitDTO Double alanlar BigDecimal'a çevrildi (hassas para/ oran serileştirme)

+++ backend/src/test/java/com/raspel/erp/controller/YoneticiKokpitControllerTest.java
@@ -47,7 +47,7 @@ class YoneticiKokpitControllerTest {
                 .ay(8)
                 .gerceklesenCiro(BigDecimal.valueOf(75000))
                 .hedefCiro(BigDecimal.valueOf(100000))
-                .ciroIlerlemeYuzdesi(75.0)
+                .ciroIlerlemeYuzdesi(BigDecimal.valueOf(75.0))
                 .build();
 
         when(kokpitService.getKokpitVerileri(eq(1L), eq(2026), eq(8))).thenReturn(dto);
